+++ src/Object/Geometry/Triangle.cpp
@@ -7,17 +7,17 @@
 #include <iostream>
 #include <utility>
 
-Triangle::Triangle(const Mat3<double> &cords)
+Triangle::Triangle(const Mat3<float> &cords)
     : vert(cords), v(cords.y - cords.x), u(cords.z - cords.x) {
     shift = (cords.x + cords.y + cords.z) / 3;
 }
 
-Vec3<double> Triangle::normal(const Vec3<double>& p) const {
+Vec3<float> Triangle::normal(const Vec3<float>& p) const {
     return (v ^ u).norm();
 }
 
-std::vector<double> Triangle::get_intersect_(const Ray& ray) const {
-    Mat3<double> mat;
+std::vector<float> Triangle::get_intersect_(const Ray& ray) const {
+    Mat3<float> mat;
     if ((u ^ v) % -ray.v > 0) {
         mat = {u, v, -ray.v};
     } else {
@@ -25,7 +25,7 @@ std::vector<double> Triangle::get_intersect_(const Ray& ray) const {
     }
     auto solution = mat.solve(ray.start - vert.x);
     if (solution.has_value()) {
-        Vec3<double> inter = solution.value();
+        Vec3<float> inter = solution.value();
         if (inter.x >= 0 && inter.y >= 0 && inter.x + inter.y <= 1 && inter.z > 0) {
             return {inter.z};
         } else {
@@ -37,6 +37,6 @@ std::vector<double> Triangle::get_intersect_(const Ray& ray) const {
 };
 
 Box Triangle::AABB() const {
-    auto gvert = Mat3<double>(position) + rotation * vert;
+    auto gvert = Mat3<float>(position) + rotation * vert;
     return Box(min(gvert.x, min(gvert.y, gvert.z)), max(max(gvert.x, gvert.y), gvert.z));
 }
